main-be/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Pillow==10.4.0
SQLAlchemy==2.0.35
eventlet==0.36.1
paramiko==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytz==2024.1
scp==0.14.5
